use integer TRANSID
git-svn-id: https://svn.code.sf.net/p/moneymanagerex/code/trunk@4607 7f5d68d0-c712-0410-839f-9b116a576cc5

+++ sql/tables_v1.sql
@@ -136,7 +136,7 @@ PAYEEID integer primary key
 -- Describe SPLITTRANSACTIONS_V1
 CREATE TABLE SPLITTRANSACTIONS_V1(
 SPLITTRANSID integer primary key
-, TRANSID numeric NOT NULL
+, TRANSID integer NOT NULL
 , CATEGID integer
 , SUBCATEGID integer
 , SPLITTRANSAMOUNT numeric
